Move the PPMd test into ArchiveTests
It opens a fixture and checks what comes out, same as every other test in
that file, so it does not need a file of its own.
